refactor(components): isolate and modularize components
