--- unisala-astro/src/pages/new-story/image.astro
@@ -1,25 +1,0 @@
----
-import MainLayout from '@/layouts/main-layout.astro';
-import type { PostDraft } from '@/types/post';
-import {Toast} from '@/components/ui/toast';
-import {VisualAidPanel} from '@/components/newstory/index';
-
-const id = Astro.url.searchParams.get('id');
-const draftId = id || Date.now().toString();
-const currentUrl = Astro.url.toString();
-
-const initialPostDraft: PostDraft = {
-    title: '',
-    tags: [],
-    reviewSubCategories: '',
-    postText: '',
-};
----
-
-<MainLayout title='Create Post'>
-    <div class="flex">
-            <!-- <ImageEditor client:only /> -->
-            <VisualAidPanel client:only />
-
-    </div>
-</MainLayout>
